refacto: move the BackendBuilder trait into backends module
