+++ crates/dojo-bindgen/src/backends/unity/mod.rs
@@ -3,8 +3,8 @@ use std::collections::HashMap;
 use async_trait::async_trait;
 use cainome::parser::tokens::Token;
 
+use crate::backends::BackendBuilder;
 use crate::error::BindgenResult;
-use crate::BackendBuilder;
 
 pub struct UnityBuilder;
